refactor(ui): extract message-part registry to dedicated module
Slice 11b.1 step 2/14 of the message-part.tsx split. Moves the part
and tool registries — the foundation that every renderer plugs into
through `PART_MAPPING[...] = ...` and `registerTool({...})` — out of
the 2400-line monolith into a 75-line registry module.

New file `message-part-registry.tsx` exports:
  - `PART_MAPPING`           the mutable part-type → renderer map
  - `registerPartComponent`  setter helper
  - `Part`                   dispatcher component (looks up the
                             renderer registered for a given part type)
  - `registerTool`           tool renderer registration
  - `getTool`                renderer lookup by tool name
  - `ToolRegistry`           legacy alias object kept for the
                             BasicTool / GenericTool callsites

`message-part.tsx` keeps the same public symbols through a re-export
block so existing callers (session-turn.tsx, stories, tests) don't
have to move yet — the migration to direct imports happens in the
next slice.

Net: message-part.tsx 2459 → 2399 lines. Typecheck GREEN, 478 pass /
3 baseline fail unchanged.
